Improve performance of live view page
Restrict which pages are shown to avoid having many elements on
screen at once. This should improve performance by limiting the
number of elements in the jQuery search, and through having
less elements on the page to need updating through Vue. Also
makes use of the v-once directive on the ScriptLineViewer
component, and this seems to have reduced update/render times
from Vue :)