Fix issue with act and scene heading not showing
If the first line of a new scene/act is cut, then the act and scene
heading would not be shown due to the way this was computed. Instead,
update the logic to find the last non-cut line, and use this to
determine whether we need to show the heading or not.