Fix logic with checking for act/scene label
The logic was checking for whether the previous line was a stage direction,
however this is incorrect for this check. I think the logic was lifted
from the check for requiring the character[group] headings, where this
check is needed.

Remove the check here, though in all reality having it present wasn't
causing incorrect behaviour.