Fix issue with act and scene heading not showing (#627)
* Fix issue with act and scene heading not showing

If the first line of a new scene/act is cut, then the act and scene
heading would not be shown due to the way this was computed. Instead,
update the logic to find the last non-cut line, and use this to
determine whether we need to show the heading or not.

* Fix logic with checking for act/scene label

The logic was checking for whether the previous line was a stage direction,
however this is incorrect for this check. I think the logic was lifted
from the check for requiring the character[group] headings, where this
check is needed.

Remove the check here, though in all reality having it present wasn't
causing incorrect behaviour.